requirements.txt
Django>=4.2,<6.0
python-decouple>=3.8
stripe>=7.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.0.0
gunicorn>=21.0.0
dj-database-url>=2.0.0